refactor: move Properties.java and rename Command.java
