+++ src/main/java/com/github/craxlor/discordbot/command/ICommand.java
@@ -9,7 +9,7 @@ import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.build.CommandData;
 
-public interface Command {
+public interface ICommand {
 
     @Nonnull
     CommandData getCommandData();
